refactoring: integrate wely.ts and class.ts
